Create  Construct BST from Preorder Traversal.cpp
Problem statement:
You are given a preorder traversal of a binary search tree. Your task is to find the postorder from the preorder.



Return the root of the BST constructed from the given preorder. The driver code will then use this root to print the post-order traversal.



For example:
You are given preOrder = [10, 5, 1, 7, 40, 50], the binary search tree from the given preorder traversal is 

Hence the answer is [1, 7, 5, 50, 40, 10].